verify: test mekle multi proof
