+++ contracts/verify/package.json
@@ -24,6 +24,7 @@
     "hardhat": "^2.3.0",
     "hardhat-abi-exporter": "^2.2.1",
     "hardhat-gas-reporter": "^1.0.4",
+    "merkletreejs": "^0.2.18",
     "shx": "^0.3.3"
   }
 }
